edi_gs1_stock: inbound info based on output mixin
