+++ edi_gs1_stock/components/__init__.py
@@ -1,1 +1,2 @@
+from . import output_mixin
 from . import inbound_instruction
